Add support for RMI SSL
RMIScout first attempts cleartext communication, falling to an all-trusting SSL
connection, as it is less common. Saves users from dealing with
keystores/validation for something they're trying to exploit XD.

+++ demo/start.sh
@@ -1,7 +1,17 @@
 #!/bin/bash
 
-rmiregistry &
-PID=$!
-
-javac *.java && java -classpath ./ -Djava.rmi.server.codebase=file:$PWD Server
-kill $PID
+javac com/bishopfox/example/*.java
+if [ "$1" ]; then
+    echo "Starting SSL Server"
+    rmic -d . com.bishopfox.example.SSLServer
+    java -classpath ./ -Djava.rmi.server.codebase=file:$PWD \
+        -Djava.security.policy=resources/policy \
+        -Djavax.net.ssl.keyStore=resources/selfsigned.jks \
+        -Djavax.net.ssl.keyStorePassword=test123 \
+        com.bishopfox.example.SSLServer
+else
+    rmiregistry &
+    PID=$!
+    java -classpath ./ -Djava.rmi.server.codebase=file:$PWD com.bishopfox.example.Server
+    kill $PID
+fi
